Clean up code, and add some new constructors for stuff.
