+++ library/src/main/java/com/agentdid127/resourcepack/library/utilities/ImageConverter.java
@@ -36,7 +36,7 @@ public class ImageConverter {
             wMultiplier = wMultiplier == 0 ? 1 : wMultiplier;
             hMultiplier = hMultiplier == 0 ? 1 : hMultiplier;
         } else {
-            Logger.log("Image '" + locationIn.getFileName() + "' is not a power of 2. Converting to be so.");
+            Logger.log("Image '" + locationIn.getFileName() + "' resolution size is not a power of 2. Converting to be so.");
             newImage = new BufferedImage((int) Math.ceil(Math.log(image.getWidth()) / Math.log(2)),
                     (int) Math.ceil(Math.log(image.getHeight()) / Math.log(2)), image.getType());
             width = (int) Math.ceil(Math.log(image.getWidth()) / Math.log(2));
